backend/pyproject.toml
[project]
name = "buybuyalert-backend"
version = "0.1.0"
description = "Price tracking and alert backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "apscheduler>=3.10.0",
    "pydantic>=2.0.0",
    "aiosqlite>=0.20.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "python-dotenv>=1.0.0",
    "playwright>=1.40.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
